+++ squidasm/run/stack/run.py
@@ -118,7 +118,7 @@ def _setup_network(config: StackNetworkConfig) -> StackNetwork:
     return StackNetwork(stacks, link_prots)
 
 
-def _run(network: StackNetwork) -> List[Dict[str, Any]]:
+def _run(network: StackNetwork) -> List[List[Dict[str, Any]]]:
     """Run the protocols of a network and programs running in that network.
 
     NOTE: For now, only two nodes and a single link are supported.
@@ -145,13 +145,13 @@ def _run(network: StackNetwork) -> List[Dict[str, Any]]:
 
 def run(
     config: StackNetworkConfig, programs: Dict[str, Program], num_times: int = 1
-) -> List[Dict[str, Any]]:
+) -> List[List[Dict[str, Any]]]:
     """Run programs on a network specified by a network configuration.
 
     :param config: configuration of the network
     :param programs: dictionary of node names to programs
     :param num_times: numbers of times to run the programs, defaults to 1
-    :return: program results
+    :return: program results, outer list is per stack, inner list is per program iteration
     """
     network = _setup_network(config)