+++ squidasm/sim/stack/netstack.py
@@ -22,6 +22,7 @@ from netsquid.components.qprogram import QuantumProgram
 from netsquid.nodes import Node
 from netsquid.qubits.ketstates import BellIndex
 from netsquid_magic.link_layer import MagicLinkLayerProtocolWithSignaling
+from netsquid_magic.magic_distributor import DoubleClickMagicDistributor
 from qlink_interface import (
     ReqCreateAndKeep,
     ReqCreateBase,
@@ -337,8 +338,14 @@ class Netstack(ComponentProtocol):
             )
             self._logger.info(f"got result for pair {pair_index}: {result}")
 
+            # This code is commented out for a hotfix as it was found that heralded link did not return
+            # Phi+ bell state. The issue needs to be investigated.
+            if isinstance(
+                self._egp._ll_prot._magic_distributor, DoubleClickMagicDistributor
+            ):
+                pass
             # Bell state corrections. Resulting state is always Phi+ (i.e. B00).
-            if result.bell_state == BellIndex.B00:
+            elif result.bell_state == BellIndex.B00:
                 pass
             elif result.bell_state == BellIndex.B01:
                 prog = QuantumProgram()